==> 02_notebooks/02_Prepare_dataset.ipynb <==
import pandas as pd
import codecs
from pandarallel import pandarallel
import datetime
import numpy as np
import feather
from pathlib import Path
import os

pd.options.display.max_columns = 999
pandarallel.initialize(progress_bar=False)


pd.options.display.max_columns = 999
pd.options.display.max_rows = 999

with codecs.open('../01_data/variables_browseNotes_select.txt','r') as f:
    variables = [i[:-1] for i in f.readlines()]

%%time

df = pd.read_csv('../01_data/loan.csv', 
                 low_memory=False,
                 usecols=variables,
                 parse_dates=['issue_d','earliest_cr_line'],
                 date_parser=lambda col: pd.to_datetime(col, format="%b-%Y", errors='coerce').date()
                )

df.dtypes

def include_exclude(col):
    if col <= pd.to_datetime('2016.03'):
        return True
    else:
        return False
def assign_random():
    return np.random.rand()

df = df.assign(include = lambda x: x['issue_d'].parallel_apply(include_exclude))

from collections import Counter
Counter(df.include.tolist())

np.random.seed(seed=666)
df['random_value'] = np.random.rand(df.shape[0])

default_map = {
    'Fully Paid': 0,
    'Default' : 1,
    'In Grace Period': 0,
    'Does not meet the credit policy. Status:Fully Paid' : 0,
    'Charged Off': 1,
    'Late (31-120 days)' : 1,
    'Current' : 0,
    'Does not meet the credit policy. Status:Charged Off' : 1,
    'Late (16-30 days)' : 0
}

def default_flag(value):
    return default_map[value]

df = df.assign(default = lambda x: x['loan_status'].parallel_apply(default_flag))

df.head()

Save excluded data

feather.write_dataframe(df.query('include == False').drop('include', axis=1), '../01_data/DS_loans_OUT.feather')


Included data

df_IN = df.query('include == True').drop('include', axis=1)

Drop irrelevant features

df_IN.select_dtypes(include=['object']).describe().T

df_IN.select_dtypes(exclude=['object']).describe().T

drop_columns_all_nan = ['id','member_id','url','revol_bal_joint','sec_app_inq_last_6mths','sec_app_mort_acc',
                        'sec_app_open_acc','sec_app_revol_util','sec_app_open_act_il','sec_app_num_rev_accts',
                        'sec_app_chargeoff_within_12_mths','sec_app_collections_12_mths_ex_med',
                        'sec_app_mths_since_last_major_derog']

df_IN =df_IN.drop(drop_columns_all_nan, axis=1)

df.shape

df_IN.shape

feather.write_dataframe(df_IN.query('random_value < 0.7'), '../01_data/DS_loans_IN_train.feather')
feather.write_dataframe(df_IN.query('random_value >= 0.7'), '../01_data/DS_loans_IN_test.feather')
